feat : implement player anomaly
