--- Assets/DesignAssets/Player/Scripts.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2cf75f156f1897f45a7ebeb5c5ff38bf
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
