--- Assets/DesignAssets/Player/Scripts/AnomalyFalseWakeupSceneController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d60a85b1a8f7ff14c86bc753f1a4a45a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
