--- Assets/DesignAssets/Player/Scripts/AnomalyTrueWakeupSceneController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b6e60feff7f56cc4bb90ade057d43205
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
