--- Assets/DesignAssets/Player/Scripts/Stage0AnimationController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6082291abd411da408f042dc9852518e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
