--- Assets/DesignAssets/Player/Scripts/Stage0AnimationPlayerController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f68545e1d0890784984976a611301b2f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
