+++ Assets/Prefabs/Player_Awake.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 0110d97a193db4843ace63fbd15777dd
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
