+++ Assets/Prefabs/Player_NotSleeping.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 98de16fd0c9804001a66babd45024c47
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
