+++ Assets/Prefabs/Player_Transparent.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 05da103948bab41a8884efaf73c72afe
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
