+++ Assets/Scenes/AnomalyFalseWakeupScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 592d78fa8b7ed4b97adb654873fe2c52
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
