+++ Assets/Scenes/AnomalyScene_LightSettings.lighting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 49d19ace89d95475d86e35b321cc7837
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 4890085278179872738
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
