+++ Assets/Scenes/AnomalyTrueWakeupScene.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 3f97b10dc16af4886843dd504692d965
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
