+++ Assets/Scenes/Stage0Animation.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8409b5fd01fff41139a0178a585630c4
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
