+++ Assets/Scripts/Animation.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ab58ec2e8295d415ba1ab8637034bd42
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
