+++ Assets/Scripts/Animation/AnomalyFalseWakeupSceneController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 650c415289fb7439983a160f6ba98e79
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
