+++ Assets/Scripts/Animation/AnomalyTrueWakeupSceneController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bd575c7e114d64fbe9c4e0fa9a5a3a81
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
