+++ Assets/Scripts/Animation/Stage0AnimationController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c83ef54dd7d0e491984de65a1c22bacf
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
