+++ Assets/Scripts/Anomaly/EasyPlayerAnomoly.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 590de1218554f49dabd666f4a5713bc2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
